Fast exits read/write more than MAX_INT KiB per day.
For example, see "other" entry in:

  exit-kibibytes-read 80=505190490,182=25102395,443=61873906,
    6881=47999666,8989=8657674,17173=7910494,21762=9138992,
    45682=5154543,50500=6086469,51413=62394452,other=2282907805

+++ src/org/torproject/descriptor/ExtraInfoDescriptor.java
@@ -205,18 +205,18 @@ public interface ExtraInfoDescriptor extends Descriptor {
   /* Return statistics on KiB written by port with map keys being ports
    * (or "other") and map values being KiB rounded up to the next full
    * KiB, or null if no exit statistics are included. */
-  public SortedMap<String, Integer> getExitKibibytesWritten();
+  public SortedMap<String, Long> getExitKibibytesWritten();
 
   /* Return statistics on KiB read by port with map keys being ports (or
    * "other") and map values being KiB rounded up to the next full KiB, or
    * null if no exit statistics are included. */
-  public SortedMap<String, Integer> getExitKibibytesRead();
+  public SortedMap<String, Long> getExitKibibytesRead();
 
   /* Return statistics on opened exit streams with map keys being ports
    * (or "other") and map values being the number of opened streams,
    * rounded up to the nearest multiple of 4, or null if no exit
    * statistics are included. */
-  public SortedMap<String, Integer> getExitStreamsOpened();
+  public SortedMap<String, Long> getExitStreamsOpened();
 
   /* Return the start of the included geoip statistics, or -1 if no geoip
    * statistics are included. */
